Allow version_compare to compare real values rather than (float) vals
Casting as a float butchers the version numbers. Examples:

* `(float) '3.10'` becomes `3.1`
* `(float) '3.9.3'` becomes `3.9`
* `(float) '3.9a'` becomes `3.9`

Right now, if someone has upgraded TEC to 3.10 and they have Pro at 3.9, the version check sees that 3.9 is greater than 3.10 (float value comparison: `3.9 > 3.1`).

PHP's `version_compare` function supports the version numbers that we tend to use, so let's not cast. This fix will allow the outdated-plugin notification message to display once again.

See: https://central.tri.be/issues/37189

+++ src/Tribe/Main.php
@@ -653,8 +653,8 @@ if ( ! class_exists( 'Tribe__Events__Main' ) ) {
 				}
 
 				// check if addons are at an older minor version
-				$addon_minor_version = (float) $plugin['current_version'];
-				$tec_minor_version   = (float) self::VERSION;
+				$addon_minor_version = $plugin['current_version'];
+				$tec_minor_version   = self::VERSION;
 				if ( version_compare( $addon_minor_version, $tec_minor_version, '<' ) ) {
 					$out_of_date_addons[] = $plugin['plugin_name'] . ' ' . $plugin['current_version'];
 				}
